chore: fixes after review
